% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/add.R
\docType{methods}
\name{add,GatingHierarchy,multipleFilterResult-method}
\alias{add,GatingHierarchy,logicalFilterResult-method}
\alias{add,GatingHierarchy,multipleFilterResult-method}
\alias{add,GatingHierarchy,ocRectRefGate-method}
\alias{add,GatingHierarchy,ocRectangleGate-method}
\title{bypass the default flowWorkspace:::.addGate}
\usage{
\S4method{add}{GatingHierarchy,multipleFilterResult}(wf, action, name = NULL,
  ...)

\S4method{add}{GatingHierarchy,logicalFilterResult}(wf, action, parent, name,
  recompute, ...)

\S4method{add}{GatingHierarchy,ocRectangleGate}(wf, action, recompute, ...)

\S4method{add}{GatingHierarchy,ocRectRefGate}(wf, action, recompute, ...)
}
\arguments{
\item{wf}{\code{GatingHierarchy} see \link[flowWorkspace]{add} in \code{flowWorkspace} package}

\item{action}{\code{ocRectangleGate} or \code{logicalFilterResult}}

\item{name}{\code{character} vector that specifies the names of populations generated by \code{multipleFilterResult}}

\item{...}{see \link[flowWorkspace]{add} in \code{flowWorkspace} package}

\item{parent}{\code{character} parent node path (full or partial)}

\item{recompute}{\code{logical} see \link[flowWorkspace]{add} in \code{flowWorkspace} package}
}
\description{
to support adding gate along with indices without loading flow data and computing

to support adding rectangleGate yet gating through boolean operations
without loading flow data
}
\details{
however it is proven that logical indices are too big to be efficiently passed around
}

