% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.table.R
\name{as.data.table}
\alias{as.data.table}
\alias{as.data.table.gatingTemplate}
\title{convert a gatingTemplate object to a data.table}
\usage{
\method{as.data.table}{gatingTemplate}(x, keep.rownames = FALSE)
}
\arguments{
\item{x}{gatingTemplate object}

\item{keep.rownames}{not used}
}
\value{
a data.table
}
\description{
It is the inverse function of gatingTemplate constructor.
}
