% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exaustive_gating.R
\name{best.separation}
\alias{best.separation}
\title{Combine all three factors (peak height, valley location and height) to determine the best the peak separation}
\usage{
best.separation(flist, fr, min.percent = 0.05, debug.mode = FALSE, ...)
}
\arguments{
\item{flist}{a list of filters/gates generated by mindensity method}

\item{fr}{flowFrame}

\item{...}{passed to density function}

\item{peak.ratio}{the minimum ratio of two peak heights in order for the gate to be considered}
}
\value{
an integer as the index of the winner gate. When it returns 0, it indicates that no further gating should be proceeded
}
\description{
Combine all three factors (peak height, valley location and height) to determine the best the peak separation
}

