\name{channels2markers}
\alias{channels2markers}
\title{For the given flow frame, we look up the given markers and return the
corresponding channels.}
\usage{
  channels2markers(flow_frame, channels)
}
\arguments{
  \item{flow_frame}{object of type \code{flowFrame}}

  \item{channels}{the channels from which we obtain the
  corresponding markers}
}
\value{
  vector of markers
}
\description{
  For the given flow frame, we look up the given markers
  and return the corresponding channels.
}

