% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtMethod-methods.R
\name{dims,gtMethod-method}
\alias{dims,gtMethod-method}
\title{get gating method dimensions}
\usage{
\S4method{dims}{gtMethod}(x)
}
\arguments{
\item{x}{\code{gtMethod}}
}
\description{
get gating method dimensions
}
