% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{.addGate_fast}
\alias{.addGate_fast}
\title{fast version of add gates to gatingset (bypassing some R checks)}
\usage{
.addGate_fast(gs, filter, name = NULL, parent = "root",
  negated = FALSE)
}
\description{
used by gating_polyFunctions
}
