% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.table.R
\name{.argDeparser}
\alias{.argDeparser}
\title{deparse a list(named) of expression into a string 
inverse function of argParser}
\usage{
.argDeparser(args, split = TRUE)
}
\description{
deparse a list(named) of expression into a string 
inverse function of argParser
}
