% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{.argParser}
\alias{.argParser}
\title{gating arguments parser}
\usage{
.argParser(txt, split = TRUE)
}
\description{
parsing the arguments read from `args` columns of csv template into 
list of paired arguments
}
