% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{.between_interval}
\alias{.between_interval}
\title{Finds values of a vector in an interval}
\usage{
.between_interval(x, interval)
}
\arguments{
\item{x}{numeric vector}

\item{interval}{numeric vector of length 2}
}
\value{
numeric vector containing the values of \code{x} between
\code{interval}. If no values are found, \code{NA} is returned.
}
\description{
Finds values of a vector in an interval
}
\examples{
z <- seq.int(1, 9, by = 2)
.between_interval(z, interval = c(2, 8))
}
