% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-functions.R
\name{.boundary}
\alias{.boundary}
\title{boundary gating function}
\usage{
.boundary(fr, pp_res = NULL, channels, min = NULL, max = NULL, ...)
}
\arguments{
\item{fr}{\code{flowFrame} object as a data input}

\item{pp_res}{not used}

\item{min, max}{the range input for constructing the \code{rectangleGate}}

\item{...}{other arguments (not used.)}
}
\value{
a \code{filter} object
}
\description{
It is useful for filtering out very large signal (e.g. FSC,SSC, Time)
}
\details{
It essentially constructs a rectangle gate from input range (min, max)
}
