% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pluginFramework.R
\name{.checkFormals}
\alias{.checkFormals}
\title{Check the formal arguments of a gating function.}
\usage{
.checkFormals(frmls = NA, type = c("gating", "preprocessing"))
}
\description{
The formal arguments need to match a certain template
We check that they do or do not.
}
