% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-functions.R
\name{.cytokine}
\alias{.cytokine}
\title{wrapper for cytokine}
\usage{
.cytokine(fr, pp_res, ...)
}
\arguments{
\item{fr}{\code{flowFrame} object as a data input}

\item{pp_res}{preprocessing result produced by the \code{preprocessing} method}

\item{...}{arguments to be passed to \link{cytokine}}
}
\value{
a \code{filter} object
}
\description{
It does some parameter preprocessing before calling the cytokine
}
