% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gating-functions.R
\name{.cytokine_cutpoint}
\alias{.cytokine_cutpoint}
\title{Constructs a cutpoint for a flowFrame by using a derivative of the kernel
density estimate}
\usage{
.cytokine_cutpoint(x, num_peaks = 1, ref_peak = 1,
  method = c("first_deriv", "second_deriv"), tol = 0.01, adjust = 1,
  side = "right", strict = TRUE, plot = FALSE, auto_tol = FALSE,
  ...)
}
\arguments{
\item{x}{a \code{numeric} vector used as input data}

\item{num_peaks}{the number of peaks expected to see. This effectively removes
any peaks that are artifacts of smoothing}

\item{ref_peak}{After \code{num_peaks} are found, this argument provides the
index of the reference population from which a gate will be obtained. By
default, the peak farthest to the left is used.}

\item{method}{the method used to select the cutpoint. See details.}

\item{tol}{the tolerance value}

\item{adjust}{the scaling adjustment applied to the bandwidth used in the
first derivative of the kernel density estimate}

\item{side}{On which side of the density do we want to gate the tail, the
\code{'right'} (default) or \code{'left'}?}

\item{strict}{\code{logical} when the actual number of peaks detected is less than \code{ref_peak}. 
an error is reported by default. But if \code{strict} is set to FALSE, then the reference peak will be reset to the peak of the far right.}

\item{auto_tol}{when TRUE, it tries to set the tolerance automatically.}

\item{...}{additional arguments passed to \code{\link{.deriv_density}}}
}
\value{
the cutpoint along the x-axis
}
\description{
We determine a gating cutpoint using either the first or second derivative of
the kernel density estimate (KDE) of the \code{x}.
}
\details{
By default, we compute the first derivative of the kernel density estimate. 
Next, we determine the lowest valley from the derivative, which corresponds to the
density's mode for cytokines. We then contruct a gating cutpoint as the value
less than the tolerance value \code{tol} in magnitude and is also greater
than the lowest valley.

Alternatively, if the \code{method} is selected as \code{second_deriv}, we
select a cutpoint from the second derivative of the KDE. Specifically, we
choose the cutpoint as the largest peak of the second derivative of the KDE
density which is greater than the reference peak.
}
