% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-functions.R
\name{.flowClust.1d}
\alias{.flowClust.1d}
\title{wrapper for flowClust.1d}
\usage{
.flowClust.1d(fr, pp_res = NULL, channels, ...)
}
\arguments{
\item{fr}{\code{flowFrame} object as a data input}

\item{pp_res}{preprocessing result produced by the \code{preprocessing} method}

\item{...}{arguments to be passed to \link{flowClust.1d}}

\item{xChannel}{must be \code{NA}.}

\item{yChannel}{the dimension used for gating}
}
\value{
a \code{filter} object
}
\description{
It does some parameter preprocessing before calling the flowClust.1d
}
