% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-functions.R
\name{.flowClust.2d}
\alias{.flowClust.2d}
\title{wrapper for flowClust.2d}
\usage{
.flowClust.2d(fr, pp_res = NULL, channels, ...)
}
\arguments{
\item{fr}{\code{flowFrame} object as a data input}

\item{pp_res}{preprocessing result produced by the \code{preprocessing} method}

\item{...}{arguments to be passed to \link{flowClust.2d}}
}
\value{
a \code{filter} object
}
\description{
It does some parameter preprocessing before calling the flowClust.2d
}
