% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{.gateToFilterResult}
\alias{.gateToFilterResult}
\title{convert gate to a filterResult}
\usage{
.gateToFilterResult(fr, channel, gate, positive)
}
\arguments{
\item{fr}{flowFrame}

\item{channel}{\code{character} channel to used for gate}

\item{gate}{\code{filter} object}

\item{positive}{\code{logical}}
}
\description{
used for computing the gate indices on the fly
}
