% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-functions.R
\name{.gating_adaptor}
\alias{.gating_adaptor}
\title{An adapter to connect the gating wrapper function with the \link{gating} method}
\usage{
.gating_adaptor(fs, pp_res, gFunc, popAlias, channels, gFunc_args)
}
\arguments{
\item{fs}{\code{flowSet} or \code{ncdfFlowSet} object}

\item{pp_res}{preprocessing result produced by the \code{preprocessing} method}

\item{gFunc}{\code{character} function name of the wrapper function to be invoked}

\item{popAlias}{\code{character} the population names that are used to determine how many gates to be expected from the gating function}

\item{gFunc_args}{arguments to be passed to wrapper function('gFunc')}
}
\value{
a \code{list} of \code{filter}s
}
\description{
It coerce the input (\code{flowSet}) to a single \code{flowFrame} and apply the gating wrapper function
then replicate the gates across samples from the \code{flowSet}.
}
