% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gating-methods.R
\name{.gating_boolMethod}
\alias{.gating_boolMethod}
\title{internal function (gating_boolMethod)}
\usage{
.gating_boolMethod(x, y, gtPop, parent, ...)
}
\description{
@param y \code{GatingSet}
 @param gtPop a \code{gtPopulation} object that contains the information of the cell population that is going to be generated by this gating method.
 @param  parent \code{character} specifying the parent node within \code{GatingSet} to which the new popoulation is to be attached.
}
