% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gating-methods.R
\name{.gating_gatingTemplate}
\alias{.gating_gatingTemplate}
\title{internal function (gating_gatingTemplate)}
\usage{
.gating_gatingTemplate(x, y, env_fct = NULL, start = "root",
  stop.at = NULL, keep.helperGates = TRUE, ...)
}
\arguments{
\item{start}{a \code{character} that specifies the population (correspoding to 'alias' column in csv template) where the gating prcoess will start from. It is useful to quickly skip some gates and go directly to the target population in the testing run.}

\item{stop.at}{a \code{character} that specifies the population (correspoding to 'alias' column in csv template) where the gating prcoess will stop at.}

\item{...}{other arguments passed to the gatingMethod-specific \code{gating} methods.}
}
\description{
internal function (gating_gatingTemplate)
}
