% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gating-methods.R
\name{.gating_gtMethod}
\alias{.gating_gtMethod}
\title{internal function (gating_gtMethod)}
\usage{
.gating_gtMethod(x, y, gtPop, parent, pp_res,
  mc.cores = getOption("mc.cores", 2L), parallel_type = c("none",
  "multicore", "cluster"), cl = NULL, ...)
}
\arguments{
\item{x}{\code{gtMethod}}

\item{y}{\code{GatingSet}}

\item{gtPop}{a \code{gtPopulation} object that contains the information of the cell population that is going to be generated by this gating method.}

\item{parent}{\code{character} specifying the parent node within \code{GatingSet} to which the new popoulation is to be attached.}

\item{pp_res}{\code{list} the preprocessing results to be used by the gating method. The names of the list has to be matched to the sample names of the data.}

\item{mc.cores}{passed to \code{multicore} package for parallel computing}

\item{parallel_type}{\code{character} specifying the parallel type. The valid options are "none", "multicore", "cluster".}

\item{cl}{\code{cluster} object passed to \code{parallel} package}
}
\description{
It is a generic gating function that does:
1. parse the gating parameters
2. group the data when applicable 
3. apply parallelism when applicable  
4. pass the flow data(maybe grouped) and preprocessing results to the adaptor function ".gating_adaptor"
5. collect the gates and add to GatingSet object
}
