% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvTemplate-parser.R
\name{.gen_1dgate}
\alias{.gen_1dgate}
\title{convert to 1d gating based on the population pattern (A+/-B+/-)}
\usage{
.gen_1dgate(this_row, strict = TRUE)
}
\description{
convert to 1d gating based on the population pattern (A+/-B+/-)
}
