% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvTemplate-parser.R
\name{.gen_dummy_ref_gate}
\alias{.gen_dummy_ref_gate}
\title{generate some dummy rows that just serves as reference without any gating method associated}
\usage{
.gen_dummy_ref_gate(this_row)
}
\description{
generate some dummy rows that just serves as reference without any gating method associated
}
