% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvTemplate-parser.R
\name{.gen_refGate}
\alias{.gen_refGate}
\title{generate reference gate based on the splitted population patterns(A+/-B+/-)}
\usage{
.gen_refGate(new_pops, this_row, ref_nodes = NULL, alias,
  strict = TRUE)
}
\description{
generate reference gate based on the splitted population patterns(A+/-B+/-)
}
