% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{.getEllipse}
\alias{.getEllipse}
\title{Creates a matrix of points on an ellipse from a fitted flowClust model.}
\usage{
.getEllipse(filter = NULL, include = seq_len(filter@K), ecol = 1,
  elty = 1, quantile = NULL, npoints = 50, subset = c(1, 2), ...)
}
\arguments{
\item{filter}{object containing the fitted \code{flowClust} model.}

\item{include}{the mixture component in the fitted \code{flowClust} model for
which the contour (ellipse) is returned}

\item{ecol}{\code{numeric} to be documented}

\item{elty}{\code{numeric} to be documented}

\item{quantile}{the contour level of the ellipse. See details.}

\item{npoints}{the number of points on the ellipse}

\item{subset}{the dimensions of the mixture component to return}

\item{\code{...}}{additional parameters}
}
\value{
matrix containing the points of the ellipse from the flowClust contour
}
\description{
The ellipse is constructed from a contour from the fitted flowClust model.
}
\details{
By default, the contour level is extracted from the \code{ruleOutliers} slot
in the \code{filter}. The user can override this level by specifying value
between 0 and 1 in \code{quantile}.
}
