% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{.getEllipseGate}
\alias{.getEllipseGate}
\title{revised based on .getEllipse.
try to construct ellipsoidGate directly from tmixFilter result instead of polygon fitting
when trans = 0}
\usage{
.getEllipseGate(filter = NULL, include = seq_len(filter@K), ecol = 1,
  elty = 1, quantile = NULL, npoints = 50, subset = c(1, 2), ...)
}
\description{
revised based on .getEllipse.
try to construct ellipsoidGate directly from tmixFilter result instead of polygon fitting
when trans = 0
}
