% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvTemplate-parser.R
\name{.getFullPath}
\alias{.getFullPath}
\title{prepend all ancester nodes to construct the full path for the given node}
\usage{
.getFullPath(ref_node, dt)
}
\arguments{
\item{ref_node}{\code{character} the node to work with}

\item{dt}{\code{data.table} that stores the current preprocessed template}
}
\description{
prepend all ancester nodes to construct the full path for the given node
}
