% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pluginFramework.R
\name{.getPluginMethods}
\alias{.getPluginMethods}
\title{return a list of registered and default gating methods}
\usage{
.getPluginMethods(x)
}
\description{
return a list of registered and default gating methods
}
