% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindensity2.R
\name{.improvedMindensity}
\alias{.improvedMindensity}
\title{a function to find interesting features in density (most notably the minimum)
as precursor to creating a 1D gate}
\usage{
.improvedMindensity(D, adjust = 2, gate_range = NA, plot = FALSE,
  ...)
}
\arguments{
\item{D}{a \code{density} containing the data to operate on}

\item{adjust}{a \code{numeric} specifying the amount of smoothing to be used}

\item{gate_range}{a \code{character} specifying the data range to operate on}

\item{plot}{a \code{boolean} specifying whether to output a plot}
}
\description{
a function to find interesting features in density (most notably the minimum)
as precursor to creating a 1D gate
}
\author{
Greg Finak, Phu T. Van
}
