% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-functions.R
\name{.mindensity}
\alias{.mindensity}
\title{wrapper for mindensity}
\usage{
.mindensity(fr, pp_res = NULL, channels, ...)
}
\arguments{
\item{fr}{\code{flowFrame} object as a data input}

\item{pp_res}{preprocessing result produced by the \code{preprocessing} method}

\item{...}{arguments to be passed to \link{mindensity}}
}
\value{
a \code{filter} object
}
\description{
It does some parameter preprocessing before calling the mindensity
}
