% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindensity2.R
\name{.mindensity2}
\alias{.mindensity2}
\title{wrapper for mindensity2}
\usage{
.mindensity2(fr, pp_res, channels, ...)
}
\arguments{
\item{fr}{\code{flowFrame} object as a data input}

\item{pp_res}{preprocessing result produced by the \code{preprocessing} method}

\item{...}{arguments to be passed to \link{mindensity}}
}
\value{
a \code{filter} object
}
\description{
It does some parameter preprocessing before calling the mindensity
}
