% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pluginFramework.R
\docType{data}
\name{.openCyto_plugin_method_lookup}
\alias{.openCyto_plugin_method_lookup}
\title{The environment holding the names of registered methods}
\format{An object of class \code{environment} of length 1.}
\usage{
.openCyto_plugin_method_lookup
}
\description{
The environment holding the names of registered methods
}
\keyword{datasets}
