% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvTemplate-parser.R
\name{.preprocess_csv}
\alias{.preprocess_csv}
\title{preprocess the csv template}
\usage{
.preprocess_csv(dt, strict = TRUE)
}
\arguments{
\item{dt}{\code{data.table} loaded directly from csv gating template}

\item{strict}{whether validity check for pop alias. It is turned on in the normal template parsing. when used with add_pop it is turned off to bypass the check on some existing boolean gates (that has ! : symbols).}
}
\value{
a preprocessed(expanded when applicable) \code{data.frame}
}
\description{
It parses the data table sequentially and does the valdidity checking and expansion row by row.
 
It expands the definition of gates or construct reference gates when necessary
}
