% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvTemplate-parser.R
\name{.preprocess_row}
\alias{.preprocess_row}
\title{The actual preprocessing logic for each entry of template}
\usage{
.preprocess_row(this_row, strict = TRUE)
}
\arguments{
\item{this_row}{a single-row \code{data.table}}
}
\value{
\code{data.table}
}
\description{
Here are the major preprocessing tasks:
1. validity check for 'alias' (special character and uniqueness check)
2. validity check for the numer of parameters('dim' column)
3. dispatch 'flowClust' method to either 'flowClust.1d' or 'flowClust.2d' based on the 'pop' and 'dims' columns
4. expand the single row to multiple rows when applicable. There are basically two types of expansion:
     4.1. expand to two 1d gates and one rectangelGate when 'pop' name is defined as quadrant pattern (e.g. "A+B+")  and 'gating_method' is not "refGate"
     4.2. expand to multiple gates when 'pop' is defined with '+/-' (e.g. "A+/-" or "A+/-B+/-")
}
