% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing-method.R
\name{.preprocessing}
\alias{.preprocessing}
\title{internal function (preprocessing)}
\usage{
.preprocessing(x, y, gtPop, parent, gm, mc.cores = 1,
  parallel_type = c("none", "multicore", "cluster"), cl = NULL, ...)
}
\arguments{
\item{x}{\code{gtMethod}}

\item{y}{\code{GatingSet}}

\item{gtPop}{a \code{gtPopulation} object that contains the information of the cell population that is going to be generated by this gating method.}

\item{parent}{\code{character} specifying the parent node within \code{GatingSet} to which the new popoulation is to be attached.}

\item{mc.cores}{passed to \code{multicore} package for parallel computing}

\item{parallel_type}{\code{character} specifying the parallel type. The valid options are "none", "multicore", "cluster".}

\item{cl}{\code{cluster} object passed to \code{parallel} package}

\item{gm:}{\code{gtMethod} object}
}
\description{
internal function (preprocessing)
}
