% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-functions.R
\name{.prior_flowClust}
\alias{.prior_flowClust}
\title{wrapper for prior_flowClust}
\usage{
.prior_flowClust(fs, gs, gm, channels, groupBy, isCollapse,
  prior_source = NULL, K = NULL, neg, pos, min, max, ...)
}
\arguments{
\item{fs}{\code{flowSet} or \code{ncdfFlowSet} object}

\item{gs}{\code{GatingSet}}

\item{gm}{\code{gtMethod}}

\item{prior_source}{\code{character} specifying the ancester node from where the prior is elicited.}

\item{K}{the number of mixture components to identify. By default, this value
is \code{NULL} and determined automatically}

\item{neg, pos}{\code{numeric} specifying how many peaks are expected on positive and negative sides from 1d density profile}

\item{min}{a numeric value that sets the lower bound for data filtering. If
\code{NULL} (default), no truncation is applied.}

\item{max}{a numeric value that sets the upper bound for data filtering. If
\code{NULL} (default), no truncation is applied.}

\item{xChannel, yChannel}{\code{character} specifying the dimensions of flow data used by \code{prior_flowClust}}
}
\value{
a \code{list} of priors, see \link{prior_flowClust} for more details
}
\description{
This wrapper does some parameter preprocessing before calls \link{prior_flowClust}
}
