% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{.quantile_flowClust}
\alias{.quantile_flowClust}
\title{Computes the quantile from flowClust for a given vector of probabilties}
\usage{
.quantile_flowClust(p, object, interval, ...)
}
\arguments{
\item{p}{vector of probabilities}

\item{object}{an object containing the \code{flowClust} fit}

\item{interval}{a vector of length 2 containing the end-points of the interval
of values to find the quantile}

\item{...}{Additional arguments that are passed to \code{uniroot} to find the
quantile.}
}
\value{
the quantile corresponding to the specified probabilities
}
\description{
We estimate the quantile from a \code{flowClust} fit with a combination of
numerical integration and a root-finding method. We are effectively
estimating the cumulative distribution function (CDF) of the mixture density
estimated by \code{flowClust}.
}
\details{
Because we are using numerical methods, we also need an \code{interval} of
values in which we will attempt to find the specified quantile.
}
