% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-functions.R
\name{.quantileGate}
\alias{.quantileGate}
\title{wrapper for quantileGate}
\usage{
.quantileGate(fr, pp_res = NULL, channels, ...)
}
\arguments{
\item{fr}{\code{flowFrame} object as a data input}

\item{pp_res}{preprocessing result produced by the \code{preprocessing} method}

\item{...}{arguments to be passed to \link{quantileGate}}
}
\value{
a \code{filter} object
}
\description{
It does some parameter preprocessing before calling the quantileGate
}
