% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-functions.R
\name{.rangeGate}
\alias{.rangeGate}
\title{wrapper for rangeGate (deprecated)}
\usage{
.rangeGate(fr, pp_res = NULL, channels, ...)
}
\arguments{
\item{fr}{\code{flowFrame} object as a data input}

\item{pp_res}{not used}

\item{...}{arguments to be passed to \link[flowStats:rangeGate]{rangeGate}}
}
\value{
a \code{filter} object
}
\description{
It does some parameter preprocessing before calling the rangeGate
}
