% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{.read.FCS.csv}
\alias{.read.FCS.csv}
\title{reading flowFrame from csv spreadsheet.It is for internal usage.}
\usage{
.read.FCS.csv(file, stains = NA)
}
\arguments{
\item{file}{\code{character} csv file name that contains the fluorescence intensities}

\item{stains}{\code{character} a vector specifying the stains for channels. Default is \code{NA}}
}
\value{
a \code{flowFrame} object
}
\description{
reading flowFrame from csv spreadsheet.It is for internal usage.
}
