% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{.read.flowSet.csv}
\alias{.read.flowSet.csv}
\title{reading flowSet from multiple csv spreadsheets.It is for internal usage.}
\usage{
.read.flowSet.csv(files, ...)
}
\arguments{
\item{files}{\code{character} csv file names}

\item{...}{arguments passed to \link{read.FCS.csv}}
}
\value{
a \code{flowSet} object
}
\description{
reading flowSet from multiple csv spreadsheets.It is for internal usage.
}
