% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pluginFramework.R
\name{.register}
\alias{.register}
\title{Register a gating function for OpenCyto}
\usage{
.register(fun = NA, methodName, type = c("gating", "preprocessing"))
}
\description{
Register a gating function for OpenCyto
}
