% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{.rotation_matrix}
\alias{.rotation_matrix}
\title{Constructs a 2x2 rotation matrix for a given angle}
\usage{
.rotation_matrix(theta)
}
\arguments{
\item{theta}{\code{numeric} the degree of rotation that ensures the angle between the x-axis and the eigenvector is between 0 and pi}
}
\description{
Constructs a 2x2 rotation matrix for a given angle
}
