% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{.scale_huber}
\alias{.scale_huber}
\title{Scales a vector of data using the Huber robust estimator for mean and
standard deviation}
\usage{
.scale_huber(x, center = TRUE, scale = TRUE)
}
\arguments{
\item{x}{numeric vector}

\item{center}{logical value. Should \code{x} be centered?}

\item{scale}{logical value. Should \code{x} be scaled?}
}
\value{
numeric vector containing the scaled data
}
\description{
This function is an analog to \code{\link{scale}} but using Huber robust
estimators instead of the usual sample mean and standard deviation.
}
