% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-functions.R
\name{.singletGate}
\alias{.singletGate}
\title{wrapper for \link[flowStats:singletGate]{singletGate}}
\usage{
.singletGate(fr, pp_res = NULL, channels, ...)
}
\arguments{
\item{fr}{\code{flowFrame} object as a data input}

\item{pp_res}{not used}

\item{...}{arguments to be passed to \link[flowStats:singletGate]{singletGate}}
}
\value{
a \code{filter} object
}
\description{
wrapper for \link[flowStats:singletGate]{singletGate}
}
