% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvTemplate-parser.R
\name{.splitTerms}
\alias{.splitTerms}
\title{split the population pattern into multiple population names}
\usage{
.splitTerms(pop_pat, two_pop_token, popName, dims)
}
\description{
currently only works for A+/-B+/- .
TODO:  support A+/-
}
