% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{.standardize_flowFrame}
\alias{.standardize_flowFrame}
\title{Standardizes a channel within a \code{flowSet} object using the mode of the
kernel density estimate and the Huber estimator of the standard deviation}
\usage{
.standardize_flowFrame(fr, channel, data = TRUE)
}
\arguments{
\item{channel}{the channel to standardize}

\item{data}{\code{logical} indicating whether to return the transformed flow data.}

\item{fs}{a \code{flowSet} object}
}
\value{
the \code{transformation} list, center, scale and optionally the transformed \code{flowFrame}
}
\description{
Standardizes a channel within a \code{flowSet} object using the mode of the
kernel density estimate and the Huber estimator of the standard deviation
}
