% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-functions.R
\name{.standardize_flowset}
\alias{.standardize_flowset}
\title{preprocessing wrapper for .standardize_flowFrame}
\usage{
.standardize_flowset(fs, gs, gm, channels, groupBy, isCollapse, ...)
}
\arguments{
\item{fs}{a \code{flowSet} object}

\item{groupBy}{\code{character} tells whether 'groupBy' column in csv template has been set}

\item{isCollapse}{\code{logical} indicates whether the gating is done on collapsed data. If so, the prior does not need to be replicated across samples.}

\item{xChannel}{not used}

\item{yChannel}{the channel to standardize}
}
\description{
preprocessing wrapper for .standardize_flowFrame
}
