% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-functions.R
\name{.tailgate}
\alias{.tailgate}
\title{tailgate}
\usage{
.tailgate(fr, pp_res = NULL, channels, ...)
}
\arguments{
\item{fr}{\code{flowFrame} object as a data input}

\item{pp_res}{preprocessing result produced by the \code{preprocessing} method}

\item{...}{arguments to be passed to \link{tailgate}}
}
\value{
a \code{filter} object
}
\description{
tailgate
}
