% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvTemplate-parser.R
\name{.unique_check_alias}
\alias{.unique_check_alias}
\title{check if alias is unique within template}
\usage{
.unique_check_alias(new_dt, alias, this_parent)
}
\arguments{
\item{new_dt}{\code{data.tab;e} contains the preprocessed rows}

\item{alias}{\code{character} the alias to be checked}

\item{this_parent}{\code{character} the full gating path of the parent node}
}
\value{
NULL when pass the check, otherwise, throw the error
}
\description{
check if alias is unique within template
}
