% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{.validateFlowClustArgs}
\alias{.validateFlowClustArgs}
\title{Validate the arguments to flowClust and tell the user early if there are problems
For now we make sure that the K argument to flowClust and to prior_flowClust are present and agree.}
\usage{
.validateFlowClustArgs(df)
}
\arguments{
\item{df}{a \code{data.frame} of the template}
}
\value{
a \code{data.frame}
}
\description{
Validate the arguments to flowClust and tell the user early if there are problems
For now we make sure that the K argument to flowClust and to prior_flowClust are present and agree.
}
