% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-functions.R
\name{.warpSet}
\alias{.warpSet}
\title{wrapper for \link[flowStats:warpSet]{warpSet}}
\usage{
.warpSet(fs, gs, gm, channels, groupBy, isCollapse, stains, ...)
}
\arguments{
\item{fs}{\code{flowSet} or \code{ncdfFlowSet} object}

\item{gs}{\code{GatingSet}}

\item{gm}{\code{gtMethod}}

\item{stains}{\code{character} passed to \link[flowStats:warpSet]{warpSet}}
}
\value{
\code{NULL}
}
\description{
wrapper for \link[flowStats:warpSet]{warpSet}
}
