\name{dummyMethod}
\alias{dummyMethod}
\title{A class to represent a dummy gating method that does nothing but serves as reference to be refered by other population}
\description{
  It is generated automatically by the csv template
  preprocessing to handle the gating function that returns
  multiple gates.
}

