% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_rlm.R
\name{fast_rlm}
\alias{fast_rlm}
\title{robust linear model using an M estimator}
\usage{
fast_rlm(x, y, maxit = 20)
}
\arguments{
\item{x}{matrix with first column as weight (default can be 1s), the rest columns are predict variable}

\item{y}{numeric vector as response}

\item{maxit}{maximum iterations}
}
\description{
rewritten in c++, till eval stats::lm.wfit r function in underlying cpp11 code 
It is internally used for singletGate, thus its output format may not be generic enough for common model fitting .
e.g. it doesn't take formula as input
}
