% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{fcEllipsoidGate-class}
\alias{fcEllipsoidGate-class}
\title{a concrete class that reprents the ellipsoidGate generated by flowClust}
\description{
It stores priors and posteriors as well as the actual ellipsoidGate.
}
