\name{fcFilterList}
\alias{fcFilterList}
\title{a class that extends \code{filterList} class.}
\usage{
  fcFilterList(x)
}
\arguments{
  \item{x}{\code{list} of \code{fcFilter} (i.e.
  \code{fcPolygonGate} or \code{fcRectangleGate})}
}
\description{
  Each filter in the filterList must extends the
  \code{fcFilter} class

  constuctor for \code{fcFilterList}
}

