% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{fcPolygonGate-class}
\alias{fcPolygonGate-class}
\title{a concrete class that reprents the polygonGate generated by flowClust}
\description{
It stores priors and posteriors as well as the actual polygonGate.
}
