\name{fcPolygonGate}
\alias{fcPolygonGate}
\title{a concrete class that reprents the polygonGate generated by flowClust}
\usage{
  fcPolygonGate(x, priors, posts)
}
\arguments{
  \item{x}{a \code{polygonGate} object}

  \item{priors}{a \code{list} storing priors}

  \item{posts}{a \code{list} storing posteriors}
}
\description{
  It stores priors and posteriors as well as the actual
  polygonGate.

  constuctor for \code{fcPolygonGate}
}

