% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{fcRectangleGate-class}
\alias{fcRectangleGate-class}
\title{a concrete class that reprents the rectangleGate generated by flowClust}
\description{
It stores priors and posteriors as well as the actual rectangleGate.
}
