\name{fcRectangleGate}
\alias{fcRectangleGate}
\title{a concrete class that reprents the rectangleGate generated by flowClust}
\usage{
  fcRectangleGate(x, priors, posts)
}
\arguments{
  \item{x}{a \code{rectangleGate} object}

  \item{priors}{a \code{list} storing priors}

  \item{posts}{a \code{list} storing posteriors}
}
\description{
  It stores priors and posteriors as well as the actual
  rectangleGate.

  constuctor for \code{fcRectangleGate}
}

