% Generated by roxygen2 (4.0.0): do not edit by hand

\name{fcTree-class}
\alias{fcTree-class}
\title{A class to represent a flowClust tree.}
\description{
  It is a graphNEL used as a container to store priors and
  posteriors for each flowClust gate that can be visualized
  for the purpose of fine-tunning parameters for flowClust
  algorithm
}

