% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{gating,gtMethod,GatingSet-method}
\alias{gating,gtMethod,GatingSet-method}
\alias{gating,gtMethod,GatingSetList-method}
\title{apply a \link[openCyto:gtMethod-class]{gtMethod} to the \code{GatingSet}}
\usage{
\S4method{gating}{gtMethod,GatingSet}(x, y, ...)
}
\arguments{
\item{x}{\code{gtMethod}}

\item{y}{\code{GatingSet}}

\item{...}{other arguments}
}
\description{
apply a \link[openCyto:gtMethod-class]{gtMethod} to the \code{GatingSet}
}

