% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exaustive_gating.R
\docType{methods}
\name{gating,GatingSet,missing-method}
\alias{gating,GatingSet,missing-method}
\alias{gating,character,GatingSet-method}
\title{automated gating}
\usage{
\S4method{gating}{GatingSet,missing}(x, y, ...)

\S4method{gating}{character,GatingSet}(x, y, ...)
}
\arguments{
\item{gs}{a GatingSet}

\item{peak.separation}{}
}
\description{
The idea is pretty simple: start with the marker with the best separation between -/+ peaks, 
then iterate over all other markers until you can't find a good separation between peaks. 
Different methods can be used for determining good/bad separation, such as the ICL or the
heuristic method based on peak distance.
}
\examples{
dataDir <- system.file("extdata",package="flowWorkspaceData")
gs <- load_gs(list.files(dataDir, pattern = "gs_manual",full = TRUE))

#clone the gating tree strucuture of the existing manually gated gs(through flowJo)
gs2 <- clone(gs, isNew = FALSE, isEmpty = FALSE)

#erase all the manual gates and start from root
Rm("not debris", gs2)

getNodes(gs2)

gating(gs2)
# specify a particular existing node to start with
gs2 <- clone(gs, isNew = FALSE, isEmpty = FALSE)
Rm("CD3+", gs2)
gating("singlets", gs2, min.count = 2000,min.percent = 0.3)
}

