% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{getGate,fcTree,character-method}
\alias{getGate,fcTree,character-method}
\title{get gates saved in \code{fcTree}}
\usage{
\S4method{getGate}{fcTree,character}(obj, y, ...)
}
\arguments{
\item{obj}{\code{fcTree}}

\item{y}{\code{character} node name}

\item{...}{other arguments (not used)}
}
\description{
get gates saved in \code{fcTree}
}

