% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatingTemplate-methods.R
\docType{methods}
\name{getGate,gatingTemplate,character-method}
\alias{getGate,gatingTemplate,character-method}
\title{get gating method from the node}
\usage{
\S4method{getGate}{gatingTemplate,character}(obj, y, z)
}
\arguments{
\item{obj}{\code{gatingTemplate}}

\item{y}{\code{character} parent node path}

\item{z}{\code{character} child node path}
}
\description{
get gating method from the node
}
\examples{
\dontrun{
gt <- gatingTemplate(system.file("extdata/gating_template/tcell.csv",package = "openCyto"))
getNodes(gt, only.names = TRUE)
getNodes(gt, "/nonDebris")
getChildren(gt, "/nonDebris")
getGate(gt, "/nonDebris", "/nonDebris/singlets")
}
}
