% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMFI.R
\name{getMFI}
\alias{getMFI}
\alias{getMFI.GatingHierarchy}
\alias{getMFI.GatingSet}
\alias{getMFI.GatingSetList}
\title{Exact MFI from populations(or nodes) for all the markers}
\usage{
getMFI(x, ...)

\method{getMFI}{GatingSetList}(x, ...)

\method{getMFI}{GatingSet}(x, ...)

\method{getMFI}{GatingHierarchy}(x, only.leaf = TRUE, ...)
}
\arguments{
\item{x}{a GatingSet or GatingHierarchy}

\item{...}{other arguments
only.leaf default is TRUE, which only get MFI for the leaf nodes
... arguments passed to \link{getNodes} method.}
}
\value{
a data.table that contains MFI values for each marker per column along with 'pop' column and 'sample' column (when used on a 'GatingSet')
}
\description{
It calculates the MFI for each marker.
}
\examples{
\dontrun{
# get MFI of each markers for all the leaf nodes
dt <- getMFI(gs)

# remove the restriction of leaf node and cutomize the node path length
getMFI(gs, only.leaf = FALSE, path = "auto")
}      
}

