% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcTree-methods.R
\name{getNodes,fcTree-method}
\alias{getNodes,fcTree-method}
\title{get nodes from \code{fcTree}}
\usage{
\S4method{getNodes}{fcTree}(x, y)
}
\arguments{
\item{x}{\code{fcTree}}

\item{y}{\code{character} node name}
}
\description{
get nodes from \code{fcTree}
}
