% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{getParent,gatingTemplate,character-method}
\alias{getParent,gatingTemplate,character-method}
\title{get parent nodes}
\usage{
\S4method{getParent}{gatingTemplate,character}(obj, y, isRef = FALSE)
}
\arguments{
\item{obj}{\code{gatingTemplate}}

\item{y}{\code{character} child node path}

\item{isRef}{\code{logical} whether show the reference node besides the parent node}
}
\description{
get parent nodes
}
\examples{
\dontrun{
gt <- gatingTemplate(system.file("extdata/gating_template/tcell.csv",package = "openCyto"))

getNodes(gt, "/nonDebris")
getParent(gt, "/nonDebris/singlets")
}
}

