% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{groupBy,gtMethod-method}
\alias{groupBy,gtMethod-method}
\title{get the grouping variable for the gating method}
\usage{
\S4method{groupBy}{gtMethod}(object)
}
\arguments{
\item{object}{\code{gtMethod}}
}
\description{
When specified, the flow data is grouped by the grouping variable (column names in pData).
Within each group, when \code{isCollapse} is set to TRUE, the gating method is applied to the collapsed data.
Otherwise, it is done indepentently for each indiviudal sample(flowFrame).
Grouping variable is also used by \code{preprocessing method}.
}

