% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatingTemplate-methods.R
\name{gt_get_children}
\alias{gt_get_children}
\alias{getChildren}
\alias{getChildren,gatingTemplate,character-method}
\title{get children nodes}
\usage{
gt_get_children(obj, y)
}
\arguments{
\item{obj}{\code{gatingTemplate}}

\item{y}{\code{character} parent node path}
}
\description{
get children nodes
}
\examples{
\dontrun{
gt <- gatingTemplate(system.file("extdata/gating_template/tcell.csv",package = "openCyto"))

gt_get_nodes(gt, "/nonDebris")
gt_get_children(gt, "/nonDebris") 
}
}
