% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatingTemplate-methods.R
\name{gt_get_gate}
\alias{gt_get_gate}
\alias{getGate,gatingTemplate,character-method}
\title{get gating method from the node}
\usage{
gt_get_gate(obj, y, z)
}
\arguments{
\item{obj}{\code{gatingTemplate}}

\item{y}{\code{character} parent node path}

\item{z}{\code{character} child node path}
}
\description{
get gating method from the node
}
\examples{
\dontrun{
gt <- gatingTemplate(system.file("extdata/gating_template/tcell.csv",package = "openCyto"))
gt_get_nodes(gt, only.names = TRUE)
gt_get_nodes(gt, "/nonDebris")
gt_get_children(gt, "/nonDebris")
gt_get_gate(gt, "/nonDebris", "/nonDebris/singlets")
}
}
