% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatingTemplate-methods.R
\name{gt_get_parent}
\alias{gt_get_parent}
\alias{getParent}
\alias{getParent,gatingTemplate,character-method}
\title{get parent nodes}
\usage{
gt_get_parent(obj, y, isRef = FALSE)
}
\arguments{
\item{obj}{\code{gatingTemplate}}

\item{y}{\code{character} child node path}

\item{isRef}{\code{logical} whether show the reference node besides the parent node}
}
\description{
get parent nodes
}
\examples{
\dontrun{
gt <- gatingTemplate(system.file("extdata/gating_template/tcell.csv",package = "openCyto"))

gt_get_nodes(gt, "/nonDebris")
gt_get_parent(gt, "/nonDebris/singlets") 
}
}
