% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pluginFramework.R
\name{gt_list_methods}
\alias{gt_list_methods}
\alias{listgtMethods}
\title{Print a list of the registered gating methods}
\usage{
gt_list_methods()
}
\value{
Does not return anything. Prints a list of the available gating methods.
}
\description{
Print a list of the registered gating methods
}
