\name{markers2channels}
\alias{markers2channels}
\title{For the given workflow, we look up the given markers and return the
corresponding channels.}
\usage{
  markers2channels(flow_frame, markers)
}
\arguments{
  \item{flow_frame}{object of type \code{flowFrame}}

  \item{markers}{the markers from which we obtain the
  corresponding channel names}
}
\value{
  vector of channel names
}
\description{
  For the given workflow, we look up the given markers and
  return the corresponding channels.
}

