% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtPopulation-methods.R
\name{names,gtPopulation-method}
\alias{names,gtPopulation-method}
\title{get population name}
\usage{
\S4method{names}{gtPopulation}(x)
}
\arguments{
\item{x}{\code{gtPopulation} object}
}
\description{
get population name
}
