% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\docType{class}
\name{ocRectRefGate-class}
\alias{ocRectRefGate-class}
\title{special gate type that mix the rectangleGate with boolean gate}
\description{
special gate type that mix the rectangleGate with boolean gate
}
