% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ocRectRefGate}
\alias{ocRectRefGate}
\title{constructor for ocRectRefGate}
\usage{
ocRectRefGate(rectGate, boolExprs)
}
\arguments{
\item{rectGate}{\code{rectangleGate}}

\item{boolExprs}{\code{character} boolean expression of reference nodes}
}
\description{
constructor for ocRectRefGate
}

