% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pop.R, R/bayes-flowClust.R,
%   R/csvTemplate-parser.R, R/gating-functions.R, R/gating-methods.R,
%   R/gatingTemplate-methods.R, R/pluginFramework.R, R/remove_pop.R,
%   R/toggle.helperGates.R
\name{openCyto-deprecated}
\alias{openCyto-deprecated}
\title{Deprecated functions in package \pkg{openCyto}.}
\description{
\code{add_pop} --> \code{\link{gs_add_gating_method}}

\code{add_pop_init} --> \code{\link{gs_add_gating_method_init}}

\code{prior_flowClust} --> \code{\link{prior_flowclust}}

\code{templateGen} --> \code{\link{gh_generate_template}}

\code{gate_flowClust_1d} --> \code{\link{gate_flowclust_1d}}

\code{gate_flowClust_2d} --> \code{\link{gate_flowclust_2d}}

\code{quantileGate} --> \code{\link{gate_quantile}}

\code{quadGate.seq} --> \code{\link{gate_quad_sequential}}

\code{quadGate.tmix} --> \code{\link{gate_quad_tmix}}

\code{gating} --> \code{\link{gt_gating}}

\code{getNodes} --> \code{\link{gt_get_nodes}}

\code{getChildren} --> \code{\link{gt_get_children}}

\code{getParent} --> \code{\link{gt_get_parent}}

\code{getGate} --> \code{\link{gt_get_gate}}

\code{listgtMethods} --> \code{\link{gt_list_methods}}

\code{registerPlugins} --> \code{\link{register_plugins}}

\code{remove_pop} --> \code{\link{gs_remove_gating_method}}

\code{toggle.helperGates} --> \code{\link{gt_toggle_helpergates}}

\code{get.helperGates} --> \code{\link{gt_get_helpergates}}

\code{delete.helperGates} --> \code{\link{gt_delete_helpergates}}
}
