% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{openCyto.options}
\alias{openCyto.options}
\title{Some global options for openCyto
See examples for the meaning of these options and how to get/set them.}
\description{
Get/set some global options for openCyto
}
\examples{
opt <- getOption("openCyto")
#the threshold of minimum cell events required for the gating algorithm to proceed
opt[["gating"]][["minEvents"]]
#to change the threshold
opt[["gating"]][["minEvents"]] <- 100
options(openCyto = opt)

#switch off the validity check flags(Not recommended)
opt[["check.pop"]] <- FALSE
options(openCyto = opt)

}
