% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{parameters,gtMethod-method}
\alias{parameters,gtMethod-method}
\title{get parameters of the gating method/function}
\usage{
\S4method{parameters}{gtMethod}(object)
}
\arguments{
\item{object}{\code{gtMethod}}
}
\description{
get parameters of the gating method/function
}

