% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcTree-methods.R
\name{plot,fcTree,character-method}
\alias{plot,fcTree,character-method}
\title{plot the flowClust gating results}
\usage{
\S4method{plot}{fcTree,character}(x, y, channel = NULL, data = NULL, ...)
}
\arguments{
\item{x}{\code{fcTree}}

\item{y}{\code{character} node name in the \code{fcTree}}

\item{channel}{\code{character} specifying the channel.}

\item{data}{\code{GatingSet} that the \code{fcTree} is associated with}

\item{...}{other arguments}
}
\description{
This provides the priors and posteriors as well as the gates for the purpose of debugging flowClust gating algorithm
}
