% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{polyFunctions-class}
\alias{polyFunctions-class}
\title{A class to represent a polyFunctions gating method.}
\description{
It extends \code{boolMethod} class and will be expanded to multiple \code{boolMethod} object.
}
