\name{posteriors}
\alias{posteriors}
\alias{posteriors,fcFilter,ANY-method}
\alias{posteriors,fcFilter,character-method}
\title{get posteriors from a \code{fcFilter} object}
\arguments{
  \item{x}{\code{fcFilter}}

  \item{y}{\code{character} or \code{missing} that specifiy
  which channel to look for}
}
\description{
  get posteriors from a \code{fcFilter} object
}

