% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{ppMethod-class}
\alias{ppMethod-class}
\alias{ppMethod}
\title{A class to represent a preprocessing method.}
\description{
It extends \code{gtMethod} class.
}
\examples{
 \dontrun{
     gt <- gatingTemplate(system.file("extdata/gating_template/tcell.csv",package = "openCyto"))
     ppMethod(gt, '3', '4')
}     
}
