% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{preprocessing,ppMethod,GatingSet-method}
\alias{preprocessing,ppMethod,GatingSet-method}
\title{apply a \link[openCyto:ppMethod-class]{ppMethod} to the \code{GatingSet}}
\usage{
\S4method{preprocessing}{ppMethod,GatingSet}(x, y, ...)
}
\arguments{
\item{x}{\code{ppMethod}}

\item{y}{\code{GatingSet} or \code{GatingSetList}}

\item{...}{other arguments}
}
\description{
apply a \link[openCyto:ppMethod-class]{ppMethod} to the \code{GatingSet}
}

