\name{preprocessing}
\alias{preprocessing}
\alias{preprocessing,ppMethod,GatingSet-method}
\title{apply a \link{ppMethod} to the \code{GatingSet}}
\arguments{
  \item{x}{\code{ppMethod}}

  \item{x}{\code{GatingSet} ... other arguments}
}
\description{
  apply a \link{ppMethod} to the \code{GatingSet}
}

