\name{priors}
\alias{priors}
\title{get priors from a \code{fcFilter} object}
\arguments{
  \item{x}{\code{fcFilter} object}

  \item{y}{\code{character} specifying channel name. if
  \code{missing} then extract priors for all the channels}
}
\description{
  get priors from a \code{fcFilter} object
}

