% Generated by roxygen2 (4.0.0): do not edit by hand

\name{refGate-class}
\alias{refGate-class}
\title{A class to represent a reference gating method.}
\description{
  It extends \code{gtMethod} class.
}
\section{Slots}{
  \describe{ \item{refNodes}{ \code{character} specifying
  the reference nodes} }
}

