% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_pop.R
\name{remove_pop}
\alias{remove_pop}
\title{Reverse the action of gating methods applied via \code{add_pop}}
\usage{
remove_pop(GatingSet)
remove_pop(GatingSetList)
}
\arguments{
\item{gs}{The \code{GatingSet} or \code{GatingSetList} for which the most recent
\code{add_pop} call should be reversed.}
}
\description{
This function provides an easy way to remove the gates and nodes created by the most
recent call to \code{\link{add_pop}} on the specified \code{GatingSet} or \code{GatingSetList},
with a separate history being maintained for each such object. \code{remove_pop} allows 
for repeated use, effectively serving as a multi-level undo function for \code{add_pop}.
}
\examples{
\dontrun{
 # add quad gates 
 add_pop(gs, gating_method = "mindensity", dims = "CCR7,CD45RA", parent = "cd4-cd8+", pop = "CCR7+/-CD45RA+/-")
 # Remove the gates and nodes resulting from that add_pop call
 remove_pop(gs)
} 
}
\seealso{
\code{\link{add_pop}} \code{\link{add_pop_init}}
}
