% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtMethod-methods.R
\name{show,boolMethod-method}
\alias{show,boolMethod-method}
\title{show method for boolMethod}
\usage{
\S4method{show}{boolMethod}(object)
}
\arguments{
\item{object}{\code{boolMethod}}
}
\description{
show method for boolMethod
}
