% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcFilter-methods.R
\name{show,fcFilter-method}
\alias{show,fcFilter-method}
\title{show method for fcFilter}
\usage{
\S4method{show}{fcFilter}(object)
}
\arguments{
\item{object}{\code{fcFilter}
show method for fcFilter}
}
\description{
show method for fcFilter
}
