% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{show,gatingTemplate-method}
\alias{show,gatingTemplate-method}
\title{show method for gatingTemplate}
\usage{
\S4method{show}{gatingTemplate}(object)
}
\arguments{
\item{object}{\code{gatingTemplate}}
}
\description{
show method for gatingTemplate
}

