% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{show,gtMethod-method}
\alias{show,gtMethod-method}
\title{show method for gtMethod}
\usage{
\S4method{show}{gtMethod}(object)
}
\arguments{
\item{object}{\code{gtMethod}
show method for gtMethod}
}
\description{
show method for gtMethod
}

