% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tautString.R
\name{tautString}
\alias{tautString}
\alias{tautStringGate}
\title{Taut String Density Estimator Gating}
\usage{
tautString(sorted_vector, modeprior = 0)

tautStringGate(fr, channel, gate_range = NULL, min = NULL,
  max = NULL, filterId = "")
}
\arguments{
\item{sorted_vector}{\code{numeric} vector of single cell expresion from a single cytometric channel.}

\item{modeprior}{\code{numeric} scalar specifying the expected number of modes. Default 0 (autodetect). Rarely should this be
set by the user.}

\item{channel}{The channel to gate.}

\item{gate_range}{The range to look for a gate, no truncation occurs.}

\item{min}{The min range of the data to truncate the flowFrame}

\item{max}{The max range of the data to truncate the flowFrame}

\item{filterId}{The id / name of the gate.}
}
\description{
The taut string density estimator gating returns 0, 1, or more gates, depending on how many modes it identifies in the data.
}
