% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/toggle.helperGates.R
\name{toggle.helperGates}
\alias{toggle.helperGates}
\title{toggle the hidden flag of the helper gates}
\usage{
toggle.helperGates(gt, gs)
}
\arguments{
\item{gt}{gatingTemplate object}

\item{gs}{GatingSet}
}
\description{
The helper gates are defined as the referred gates in csv template. And all the chidlren of referred gates are also
referred gates thus they are considered the helper gates and can usually be hidden to simply the final gating tree.
}
\examples{
\dontrun{
gt <- gatingTemplate(gtFile)
#run the gating
gating(gt, gs)
#hide the gates that are not of interest
toggle.helperGates(gt, gs)
}
}

