% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\docType{methods}
\name{add,GatingHierarchy,ocRectangleGate-method}
\alias{add,GatingHierarchy,ocRectangleGate-method}
\alias{add,GatingHierarchy,ocRectRefGate-method}
\title{bypass the default flowWorkspace:::.addGate}
\usage{
\S4method{add}{GatingHierarchy,ocRectangleGate}(wf, action, recompute, ...)

\S4method{add}{GatingHierarchy,ocRectRefGate}(wf, action, recompute, ...)
}
\arguments{
\item{wf}{\code{GatingHierarchy} see \link[flowWorkspace]{add} in \code{flowWorkspace} package}

\item{action}{\code{ocRectangleGate} or \code{logicalFilterResult}}

\item{recompute}{\code{logical} see \link[flowWorkspace]{add} in \code{flowWorkspace} package}

\item{...}{see \link[flowWorkspace]{add} in \code{flowWorkspace} package}
}
\description{
to support adding gate along with indices without loading flow data and computing

to support adding rectangleGate yet gating through boolean operations 
without loading flow data
}
\details{
however it is proven that logical indices are too big to be efficiently passed around
}
