% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{pop_add.ocRectangleGate}
\alias{pop_add.ocRectangleGate}
\alias{pop_add.ocRectRefGate}
\title{bypass the default flowWorkspace:::.addGate}
\usage{
\method{pop_add}{ocRectangleGate}(gate, gh, recompute, ...)

\method{pop_add}{ocRectRefGate}(gate, gh, recompute, ...)
}
\arguments{
\item{gate}{\code{ocRectangleGate} or \code{logicalFilterResult}}

\item{gh}{\code{GatingHierarchy} see \link[flowWorkspace]{add} in \code{flowWorkspace} package}

\item{recompute}{\code{logical} see \link[flowWorkspace]{add} in \code{flowWorkspace} package}

\item{...}{see \link[flowWorkspace]{add} in \code{flowWorkspace} package}
}
\description{
to support adding gate along with indices without loading flow data and computing

to support adding rectangleGate yet gating through boolean operations 
without loading flow data
}
\details{
however it is proven that logical indices are too big to be efficiently passed around
}
