% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{alias,gtPopulation-method}
\alias{alias,gtPopulation-method}
\title{get population alias}
\usage{
\S4method{alias}{gtPopulation}(object)
}
\arguments{
\item{object}{\code{gtPopulation}}
}
\description{
get population alias
}

