% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{dims,gtMethod-method}
\alias{dims,gtMethod-method}
\title{get gating method dimensions}
\usage{
\S4method{dims}{gtMethod}(object)
}
\arguments{
\item{object}{\code{gtMethod}}
}
\description{
get gating method dimensions
}

