% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fcEllipsoidGate}
\alias{fcEllipsoidGate}
\title{constuctor for \code{fcEllipsoidGate}}
\usage{
fcEllipsoidGate(x, priors, posts)
}
\arguments{
\item{x}{a \code{ellipsoidGate} object}

\item{priors}{a \code{list} storing priors}

\item{posts}{a \code{list} storing posteriors}
}
\description{
constuctor for \code{fcEllipsoidGate}
}

