% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{fcFilter-class}
\alias{fcFilter-class}
\title{a virtual class that represents the gating result generated by flowClust gating function}
\description{
Bascially it extends flowCore 'filter classes to have extra slot to store priors and posteriors
}
