% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fcFilterList}
\alias{fcFilterList}
\title{constuctor for \code{fcFilterList}}
\usage{
fcFilterList(x)
}
\arguments{
\item{x}{\code{list} of \code{fcFilter} (i.e. \code{fcPolygonGate} or \code{fcRectangleGate})}
}
\description{
constuctor for \code{fcFilterList}
}

