\name{fcPolygonGate}
\alias{fcPolygonGate}
\title{constuctor for \code{fcPolygonGate}}
\usage{
  fcPolygonGate(x, priors, posts)
}
\arguments{
  \item{x}{a \code{polygonGate} object}

  \item{priors}{a \code{list} storing priors}

  \item{posts}{a \code{list} storing posteriors}
}
\description{
  constuctor for \code{fcPolygonGate}
}

