\name{fcRectangleGate}
\alias{fcRectangleGate}
\title{constuctor for \code{fcRectangleGate}}
\usage{
  fcRectangleGate(x, priors, posts)
}
\arguments{
  \item{x}{a \code{rectangleGate} object}

  \item{priors}{a \code{list} storing priors}

  \item{posts}{a \code{list} storing posteriors}
}
\description{
  constuctor for \code{fcRectangleGate}
}

