% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{fcTree}
\alias{fcTree}
\title{constructor of \code{fcTree}}
\usage{
fcTree(gt)
}
\arguments{
\item{gt}{a \code{gatingTemplate} object}
}
\description{
It adds an extra node data slot "fList"(which is a \code{filterList} object) to the \code{gatingTemplate}
}
