% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{filterObject,logicalFilterResult-method}
\alias{filterObject,logicalFilterResult-method}
\title{construct logicalGate obj}
\usage{
\S4method{filterObject}{logicalFilterResult}(x)
}
\arguments{
\item{x}{\code{logicalFilterResult}}
}
\description{
construct logicalGate obj
}

