% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gating-functions.R
\name{gate_quad_sequential}
\alias{gate_quad_sequential}
\alias{quadGate.seq}
\title{sequential quadrant gating function}
\usage{
gate_quad_sequential(fr, channels, gFunc, min = NULL, max = NULL, ...)
}
\arguments{
\item{fr}{\code{flowFrame}}

\item{channels}{\code{character} two channels used for gating}

\item{gFunc}{the name of the 1d-gating function to be used for either dimension}

\item{min}{a numeric vector that sets the lower bounds for data filtering}

\item{max}{a numeric vector that sets the upper bounds for data filtering}

\item{...}{other arguments passed to \code{.find_peak} (e.g. 'num_peaks' and 'adjust'). see \link{tailgate}}
}
\value{
a \code{filters} that contains four rectangleGates
}
\description{
The order of 1d-gating is determined so that the gates better capture the 
distributions of flow data.
}
