% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gating-functions.R
\name{gate_tail}
\alias{gate_tail}
\alias{tailgate}
\alias{cytokine}
\title{Gates the tail of a density using the derivative of a kernel density estimate}
\usage{
gate_tail(fr, channel, filterId = "", num_peaks = 1, ref_peak = 1,
  strict = TRUE, tol = 0.01, positive = TRUE, side = "right",
  min = NULL, max = NULL, bias = 0, ...)

tailgate(fr, channel, filterId = "", num_peaks = 1, ref_peak = 1,
  strict = TRUE, tol = 0.01, positive = TRUE, side = "right",
  min = NULL, max = NULL, bias = 0, ...)

cytokine(fr, channel, filterId = "", num_peaks = 1, ref_peak = 1,
  tol = 0.01, positive = TRUE, side = "right", ...)
}
\arguments{
\item{fr}{a \code{flowFrame} object}

\item{channel}{the channel from which the cytokine gate is constructed}

\item{filterId}{the name of the filter}

\item{num_peaks}{the number of peaks expected to see. This effectively removes
any peaks that are artifacts of smoothing}

\item{ref_peak}{After \code{num_peaks} are found, this argument provides the
index of the reference population from which a gate will be obtained.}

\item{strict}{\code{logical} when the actual number of peaks detected is less than \code{ref_peak}. 
an error is reported by default. But if \code{strict} is set to FALSE, then the reference peak will be reset to the peak of the far right.}

\item{tol}{the tolerance value used to construct the cytokine gate from the
derivative of the kernel density estimate}

\item{positive}{If \code{TRUE}, then the gate consists of the entire real
line to the right of the cutpoint. Otherwise, the gate is the entire real
line to the left of the cutpoint. (Default: \code{TRUE})}

\item{side}{On which side of the density do we want to gate the tail, the
\code{'right'} (default) or \code{'left'}?}

\item{min}{a numeric value that sets the lower boundary for data filtering}

\item{max}{a numeric value that sets the upper boundary for data filtering}

\item{bias}{a numeric value that adds a constant to the calculated cutpoint(threshold). Default is 0.}

\item{...}{additional arguments.}
}
\value{
a \code{filterList} containing the gates (cutpoints) for each sample
}
\description{
Gates the tail of a density using the derivative of a kernel density estimate
}
\examples{
\dontrun{
 gate <- gate_tail(fr, Channel = "APC-A") # fr is a flowFrame
}
}
