% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gating-methods.R
\docType{methods}
\name{gating,boolMethod,GatingSet-method}
\alias{gating,boolMethod,GatingSet-method}
\alias{gating,boolMethod,GatingSetList-method}
\title{apply a \code{boolMethod} to the \code{GatingSet}}
\usage{
\S4method{gating}{boolMethod,GatingSet}(x, y, ...)
}
\arguments{
\item{x}{\code{boolMethod}}

\item{...}{other arguments}
}
\description{
apply a \code{boolMethod} to the \code{GatingSet}
}

