% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gating-methods.R
\docType{methods}
\name{gating,refGate,GatingSet-method}
\alias{gating,dummyMethod,GatingSet-method}
\alias{gating,dummyMethod,GatingSetList-method}
\alias{gating,refGate,GatingSet-method}
\alias{gating,refGate,GatingSetList-method}
\title{apply a \code{refGate} to the \code{GatingSet}}
\usage{
\S4method{gating}{refGate,GatingSet}(x, y, ...)
}
\arguments{
\item{x}{\code{refGate}}

\item{y}{\code{GatingSet}}

\item{...}{other arguments}
}
\description{
apply a \code{refGate} to the \code{GatingSet}
}

