% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{gating,refGate,GatingSet-method}
\alias{gating,dummyMethod,GatingSet-method}
\alias{gating,dummyMethod,GatingSetList-method}
\alias{gating,refGate,GatingSet-method}
\alias{gating,refGate,GatingSetList-method}
\title{apply a \code{refGate} to the \code{GatingSet}}
\usage{
\S4method{gating}{refGate,GatingSet}(x, y, ...)
}
\arguments{
\item{x}{\code{refGate}}

\item{...}{other arguments}

\item{y}{\code{GatingSet}}
}
\description{
apply a \code{refGate} to the \code{GatingSet}
}

