\name{gating}
\alias{gating}
\alias{gating,boolMethod,GatingSet-method}
\alias{gating,boolMethod,GatingSetList-method}
\alias{gating,gatingTemplate,GatingSet-method}
\alias{gating,gatingTemplate,GatingSetList-method}
\alias{gating,gtMethod,GatingSet-method}
\alias{gating,gtMethod,GatingSetList-method}
\alias{gating,polyFunctions,GatingSet-method}
\alias{gating,polyFunctions,GatingSetList-method}
\alias{gating,refGate,GatingSet-method}
\alias{gating,refGate,GatingSetList-method}
\title{Applies gatingTemplate to one GatingSet.}
\arguments{
  \item{x}{a \code{gatingTemplate} object}

  \item{y}{a \code{GatingSet} object}

  \item{env_fct}{a \code{environment} that contains
  \code{fcTree} object named as 'fct'. If NULL (by
  default), no \code{fcTree} will be constructed. It is
  currently reserved for the internal debugging.}

  \item{...}{other arguments}

  \item{...}{other arguments}

  \item{x}{\code{boolMethod}}

  \item{...}{other arguments}

  \item{x}{\code{polyFunctions}}

  \item{...}{other arguments}

  \item{x}{\code{refGate}}

  \item{..}{other arguments}
}
\value{
  Nothing. As the side effect, gates generated by gating
  methods are saved in \code{GatingSet}.
}
\description{
  It loads the gating methods by topological order and
  applies them to \code{GatingSet}.

  apply a \link[openCyto:gtMethod-class]{gtMethod} to the
  \code{GatingSet}

  apply a \code{boolMethod} to the \code{GatingSet}

  It generates a batch of \code{boolMethod}s based on the
  expression defined in \code{polyFunctions} objects. It is
  a convenience way to generate different boolean
  combinations of cytokine gates.

  apply a \code{refGate} to the \code{GatingSet}
}
\examples{
\dontrun{
 gt <- gatingTemplate(file.path(path, "data/ICStemplate.csv"), "ICS")
 gs <- GatingSet(fs) #fs is a flowSet/ncdfFlowSet
 gating(gt, gs)
}
}

