% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{getNodes,fcTree-method}
\alias{getNodes,fcTree-method}
\title{get nodes from \code{fcTree}}
\usage{
\S4method{getNodes}{fcTree}(x, y)
}
\arguments{
\item{x}{\code{fcTree}}

\item{y}{\code{character} node name}
}
\description{
get nodes from \code{fcTree}
}

