% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{getNodes,gatingTemplate-method}
\alias{getNodes,gatingTemplate-method}
\title{get nodes from \link[openCyto:gatingTemplate-class]{gatingTemplate} object}
\usage{
\S4method{getNodes}{gatingTemplate}(x, y, order = c("default", "bfs", "dfs",
  "tsort"), only.names = FALSE)
}
\arguments{
\item{x}{\code{gatingTemplate}}

\item{y}{\code{character} node index. When \code{missing}, return all the nodes}

\item{order}{\code{character} specifying the order of nodes. options are "default", "bfs", "dfs", "tsort"}

\item{only.names}{\code{logical} specifiying whether user wants to get the entire \code{gtPopulation} object or just the name of the population node}
}
\description{
get nodes from \link[openCyto:gatingTemplate-class]{gatingTemplate} object
}
\examples{
\dontrun{
gt <- gatingTemplate(system.file("extdata/gating_template/tcell.csv",package = "openCyto"))
getNodes(gt)[1:2]
getNodes(gt, only.names = TRUE)
getNodes(gt, "/nonDebris")
}
}

