% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtMethod-methods.R
\name{isCollapse,gtMethod-method}
\alias{isCollapse,gtMethod-method}
\title{get the flag that determines whether gating method is applied on collapsed data}
\usage{
\S4method{isCollapse}{gtMethod}(object)
}
\arguments{
\item{object}{\code{gtMethod}}
}
\value{
\code{logical}
}
\description{
When TRUE, the flow data(multiple flowFrames) is collapsed into one and the gating method is applied on the collapsed data.
Once the gate is generated, it is thenreplicated and applied to the each single flowFrame.
}
