% Generated by roxygen2 (4.0.1): do not edit by hand
\name{listgtMethods}
\alias{listgtMethods}
\title{Print a list of the registered gating methods}
\usage{
listgtMethods()
}
\value{
Does not return anything. Prints a list of the available gating methods.
}
\description{
Print a list of the registered gating methods
}

