% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{names,gtMethod-method}
\alias{names,gtMethod-method}
\title{get gating method name}
\usage{
\S4method{names}{gtMethod}(x)
}
\arguments{
\item{x}{\code{gtMethod}}
}
\description{
get gating method name
}
\examples{
\dontrun{
gt <- gatingTemplate(system.file("extdata/gating_template/tcell.csv",package = "openCyto"))

gtMthd <- getGate(gt, "/nonDebris/singlets",  "/nonDebris/singlets/lymph")
names(gtMthd)
dims(gtMthd)
parameters(gtMthd)
isCollapse(gtMthd)
groupBy(gtMthd)

gtPop <- getNodes(gt, "/nonDebris/singlets/lymph/cd3/cd4+cd8-/CD38+")
names(gtPop)
alias(gtPop)
}
}

