% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{ocRectRefGate}
\alias{ocRectRefGate}
\title{constructor for ocRectRefGate}
\usage{
ocRectRefGate(rectGate, boolExprs)
}
\arguments{
\item{rectGate}{\code{rectangleGate}}

\item{boolExprs}{\code{character} boolean expression of reference nodes}
}
\description{
constructor for ocRectRefGate
}
