\docType{package}
\name{openCyto}
\alias{openCyto}
\alias{openCyto-package}
\title{Hierarchical Gating Pipeline for flow cytometry data}
\description{
  Hierarchical Gating Pipeline for flow cytometry data.
}
\details{
  openCyto is a package designed to facilitate the
  automated gating methods in sequential way to mimic the
  manual gating strategy.

  \tabular{ll}{ Package: \tab openCyto\cr Type: \tab
  Package\cr Version: \tab 0.99.1\cr Date: \tab
  2013-08-10\cr License: \tab GPL (>= 2)\cr LazyLoad: \tab
  yes\cr }
}
\examples{
\dontrun{gatingTemplate('test.csv')}
}
\author{
  Mike Jiang \email{wjiang2@fhcrc.org}, John Ramey
  \email{jramey@fhcrc.org}, Greg Finak
  \email{gfinak@fhcrc.org}

  Maintainer: Mike Jiang \email{wjiang2@fhcrc.org}
}
\seealso{
  See \code{\link{gating}}, \code{\link{flowClust.1d}}, for
  an overview of gating functions.
}
\keyword{package}

