% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtMethod-methods.R
\name{parameters,gtMethod-method}
\alias{parameters,gtMethod-method}
\title{get parameters of the gating method/function}
\usage{
\S4method{parameters}{gtMethod}(object)
}
\arguments{
\item{object}{\code{gtMethod}}
}
\description{
get parameters of the gating method/function
}
