% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{posteriors,fcFilter,ANY-method}
\alias{posteriors,fcFilter,ANY-method}
\alias{posteriors,fcFilter,character-method}
\title{get posteriors from a \code{fcFilter} object}
\usage{
\S4method{posteriors}{fcFilter,ANY}(x, y = "missing")
}
\arguments{
\item{x}{\code{fcFilter}}

\item{y}{\code{character} or \code{missing} that specifiy which channel to look for}
}
\description{
get posteriors from a \code{fcFilter} object
}

