% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{ppMethod-class}
\alias{ppMethod-class}
\title{A class to represent a preprocessing method.}
\description{
It extends \code{gtMethod} class.
}
\examples{
\dontrun{
     gt <- gatingTemplate(system.file("extdata/gating_template/tcell.csv",package = "openCyto"))
     ppMethod(gt, '3', '4')
}
}

