% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{priors,fcFilter,ANY-method}
\alias{priors,fcFilter,ANY-method}
\alias{priors,fcFilter,character-method}
\title{get priors from a \code{fcFilter} object}
\usage{
\S4method{priors}{fcFilter,ANY}(x, y = "missing")
}
\arguments{
\item{x}{\code{fcFilter} object}

\item{y}{\code{character} specifying channel name. if \code{missing} then extract priors for all the channels}
}
\description{
get priors from a \code{fcFilter} object
}

