% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{refGate-class}
\alias{refGate-class}
\title{A class to represent a reference gating method.}
\description{
It extends \code{gtMethod} class.
}
\section{Slots}{

\describe{ 
 \item{refNodes}{ \code{character} specifying the reference nodes}
}
}

