% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{refGate-class}
\alias{refGate-class}
\title{A class to represent a reference gating method.}
\description{
It extends \code{gtMethod} class.
}
\section{Slots}{

\describe{
 \item{refNodes}{ \code{character} specifying the reference nodes}
}
}

