% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtMethod-methods.R
\name{show,gtMethod-method}
\alias{show,gtMethod-method}
\title{show method for gtMethod}
\usage{
\S4method{show}{gtMethod}(object)
}
\arguments{
\item{object}{\code{gtMethod}
show method for gtMethod}
}
\description{
show method for gtMethod
}
