% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toggle.helperGates.R
\name{toggle.helperGates}
\alias{toggle.helperGates}
\alias{get.helperGates}
\alias{delete.helperGates}
\title{toggle/delete the hidden flag of the helper gates}
\usage{
toggle.helperGates(gt, gs)

get.helperGates(gt, gs)

delete.helperGates(gt, gs)
}
\arguments{
\item{gt}{gatingTemplate object}

\item{gs}{GatingSet}
}
\description{
The helper gates are defined as the referred gates in csv template. And all the chidlren of referred gates are also 
referred gates thus they are considered the helper gates and can usually be hidden to simply the final gating tree.
}
\details{
Note that delete action is NOT reversible.
}
\examples{
\dontrun{
gt <- gatingTemplate(gtFile)
#run the gating
gating(gt, gs)
#hide the gates that are not of interest
toggle.helperGates(gt, gs) 
#or simply remove them if you are sure they will not be useful in future
delete.helperGates(gt, gs) 
}
}
